pyproject.toml
[project]
name = "pngquant-py"
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Rust",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]
dynamic = ["version"]
dependencies = ["pngquant>=3.0.3"]
license = "GPL-3.0"
authors = [{ name = "rogerogers", email = "rogers@rogerogers.com" }]

[project.urls]
Homepage = "https://github.com/rogerogers/pngquant-py"
Repository = "https://github.com/rogerogers/pngquant-py"


[build-system]
requires = ["maturin>=1.10,<2.0"]
build-backend = "maturin"

[dependency-groups]
dev = ["maturin>=1.10.2"]
